+++ languagetool-core/src/main/java/org/languagetool/rules/LongSentenceRule.java
@@ -40,6 +40,7 @@ public class LongSentenceRule extends TextLevelRule {
   public static final String RULE_ID = "TOO_LONG_SENTENCE";
 
   private static final Pattern QUOTED_SENT_END = Pattern.compile("[?!.][\"“”„»«]", Pattern.DOTALL);
+  private static final Pattern SENT_END = Pattern.compile("[?!.]");
 
   private final ResourceBundle messages;
   private final int maxWords;
@@ -117,29 +118,26 @@ public class LongSentenceRule extends TextLevelRule {
               fromPosToken = tokens[i];
             }
             if (numWords == maxWords) {
-
               //Get last word token
               if (toPosToken == null) {
                 for (int j = tokens.length - 1; j >= 0; j--) {
                   if (isWordCount(tokens[j].getToken())) {
-                    if (tokens.length > j + 1 && tokens[j+1].getToken().equals(".")) {
+                    if (tokens.length > j + 1 && SENT_END.matcher(tokens[j+1].getToken()).matches()) {
                       toPosToken = tokens[j + 1];
                     } else {
                       toPosToken = tokens[j];
-
                     }
                     break;
                   }
                 }
               }
-
               if (fromPosToken != null && toPosToken != null) {
                 fromPos.add(fromPosToken.getStartPos());
-                toPos.add(toPosToken.getEndPos() - 1);
+                toPos.add(toPosToken.getEndPos());
               } else {
                 //keep old logic if we could not find word tokens
                 fromPos.add(tokens[0].getStartPos());
-                toPos.add(tokens[tokens.length - 1].getEndPos() - 1);
+                toPos.add(tokens[tokens.length - 1].getEndPos());
               }
               break;
             }
@@ -188,7 +186,4 @@ public class LongSentenceRule extends TextLevelRule {
   public String getConfigureText() {
     return messages.getString("guiLongSentencesText");
   }
-
-
-
 }