+++ languagetool-core/src/test/java/org/languagetool/rules/WordListValidatorTest.java
@@ -34,7 +34,7 @@ public class WordListValidatorTest {
           "[ 0-9a-zA-ZöäüÖÄÜßëçèéáàóòÈÉÁÀÓÒÍãñíîŞş&*_:\\\\" +
           "___INSERT___" +
           "Œ€ūαΑβΒγɣΓδΔεΕζΖηΗθΘιΙκΚλΛμΜνΝξΞοΟπΠρΡσΣτΤυΥφΦχΧψΨωΩάΆέΈίΊήΉύΎϊϋΰΐœţłń" +
-          "ŚśōżúïÎôêâû" +
+          "ŚśŌōżúïÎôêâû" +
           "ÇÃÕÚÊÂÔ" +
           "'’" +
           "./%-]+" +
@@ -88,6 +88,7 @@ public class WordListValidatorTest {
           "Hyndluljóð",
           "Kazanlǎk",
           "Kesäranta",
+          "Krišjānis",
           "Kŭrdzhali",
           "Malko Tŭrnovo",
           "Rígsþula",
@@ -150,6 +151,8 @@ public class WordListValidatorTest {
           "Tuğrul",
           "Kīlauea",
           "Māori",
+          "Chișinău",
+          "Chișinău/S",
           "Terzić",
           "Jūjutsu/S",
           "Ñuñoa",
@@ -169,12 +172,14 @@ public class WordListValidatorTest {
           "Ørsted",
           "Samsø/S",
           "Kılıçdaroğlu/S",
+          "Vlahović/S",
           "Sønderborg/S",
           "Sønderborg",
           "Mirotić",
           "Subotić",
           "Pÿur",
           "Subašić",
+          "Wałęsa",
           "celebrytę", // for PL
           "antybiotykoterapię", // for PL
           "elektromobilność", // for PL