+++ languagetool-http-client/src/main/java/org/languagetool/remote/SentenceAnnotator.java
@@ -150,7 +150,11 @@ public class SentenceAnnotator {
           break;
         case "i":
           fpMatches.add(getMatchIdentifier(sentence, match));
-          errorType = "IW";
+          errorType = "IM";
+          break;
+        case "b":
+          fpMatches.add(getMatchIdentifier(sentence, match));
+          errorType = "BO";
           break;
         case "f":
           fpMatches.add(getMatchIdentifier(sentence, match));
@@ -295,6 +299,7 @@ public class SentenceAnnotator {
       return sb.toString();
     }
     sb.append("(I)gnoreMatch ");
+    sb.append("(B)othOK ");
     sb.append("(F)P ");
     if (match.getReplacements().get().size() > 0) {
       sb.append("\nSUGGESTIONS: ");