+++ languagetool-server/src/test/java/org/languagetool/server/HTTPServerTest.java
@@ -231,7 +231,7 @@ public class HTTPServerTest {
     String res = check("text", "/v2/check", english, null, "A text.", "&sourceLanguage=de-DE&sourceText=Text");
     assertTrue(res.contains("DIFFERENT_PUNCTUATION"));   // bitext rule actually active
 
-    String allToneTagRulesMatchingSentence = "We have a Formal-Clarity-Match, a No-Tone-Needed-Match, a Confident-Academic-Scientific-Match, a Confident-Academic-Match match, a Picky-Clarity-Confident-Academic-Match, a Picky-Clarity-Confident-Academic-Scientific-Match, a Persuasive-Objective-Match and a Persuasive-Objective-Informal-Match.";
+    String allToneTagRulesMatchingSentence = "We have a Formal-Clarity-Match, a No-Tone-Needed-Match, a Confident-Academic-Scientific-Match, a Confident-Academic-Match match, a Picky-Clarity-Confident-Academic-Match, a Picky-Clarity-Confident-Academic-Scientific-Match, a Persuasive-Objective-Match Persuasive-Objective-Informal-Match, Persuasive-Goal-Specific-Match, and a Persuasive-Not-Goal-Specific-Match.";
 
     Language xxlang = Languages.getLanguageForShortCode("xx");
     String test1 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation");
@@ -243,6 +243,8 @@ public class HTTPServerTest {
     assertFalse(test1.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertTrue(test1.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertTrue(test1.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test1.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertTrue(test1.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
 
     String test2 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=");
     assertTrue(test2.contains("Formal_Clarity_TONE_RULE"));
@@ -253,6 +255,8 @@ public class HTTPServerTest {
     assertFalse(test2.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertTrue(test2.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertTrue(test2.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test2.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertTrue(test2.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test3 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=clarity");
     assertTrue(test3.contains("Formal_Clarity_TONE_RULE"));
@@ -263,6 +267,8 @@ public class HTTPServerTest {
     assertFalse(test3.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertFalse(test3.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertFalse(test3.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test3.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertFalse(test3.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test4 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=academic");
     assertFalse(test4.contains("Formal_Clarity_TONE_RULE")); //Clarity is enabled by default if no toneTag is specified
@@ -273,6 +279,8 @@ public class HTTPServerTest {
     assertFalse(test4.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertFalse(test4.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertFalse(test4.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test4.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertFalse(test4.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test5 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=scientific");
     assertFalse(test5.contains("Formal_Clarity_TONE_RULE"));
@@ -283,6 +291,8 @@ public class HTTPServerTest {
     assertFalse(test5.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertFalse(test5.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertFalse(test5.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test5.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertFalse(test5.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test6 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=academic&level=picky");
     assertFalse(test6.contains("Formal_Clarity_TONE_RULE"));
@@ -293,6 +303,8 @@ public class HTTPServerTest {
     assertTrue(test6.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertFalse(test6.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertFalse(test6.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test6.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertFalse(test6.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test7 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=scientific&level=picky");
     assertFalse(test7.contains("Formal_Clarity_TONE_RULE"));
@@ -303,6 +315,8 @@ public class HTTPServerTest {
     assertTrue(test7.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertFalse(test7.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertFalse(test7.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test7.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertFalse(test7.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test8 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=scientific,NO_TONE_RULE&level=picky");
     assertFalse(test8.contains("Formal_Clarity_TONE_RULE"));
@@ -313,6 +327,8 @@ public class HTTPServerTest {
     assertTrue(test8.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertFalse(test8.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertFalse(test8.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test8.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertFalse(test8.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test9 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=persuasive");
     assertFalse(test9.contains("Formal_Clarity_TONE_RULE"));
@@ -323,6 +339,8 @@ public class HTTPServerTest {
     assertFalse(test9.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertTrue(test9.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertTrue(test9.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertTrue(test9.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertTrue(test9.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
     
     String test10 = checkV2(xxlang, allToneTagRulesMatchingSentence, "&disabledRules=test_unification_with_negation&toneTags=informal");
     assertFalse(test10.contains("Formal_Clarity_TONE_RULE"));
@@ -333,6 +351,8 @@ public class HTTPServerTest {
     assertFalse(test10.contains("PICKY-CLARITY_CONFIDENT_ACADEMIC_SCIENTIFIC_TONE_RULE"));
     assertFalse(test10.contains("PERSUASIVE_OBJECTIVE_TONE_RULE"));
     assertTrue(test10.contains("PERSUASIVE_OBJECTIVE_INFORMAL_TONE_RULE"));
+    assertFalse(test10.contains("PERSUASIVE_GOAL_SPECIFIC_TONE_RULE"));
+    assertFalse(test10.contains("PERSUASIVE_NOT_GOAL_SPECIFIC_TONE_RULE"));
   }
 
   private void runDataTests() throws IOException {